Fixes issue #1.
Created a `TRegGA.config.example` file, which the user should copy to `TRegGA.config` and modify for their system setup.
Currently there is only a single variable `TRegGA_DIR` but as we review the code we should move system-specific things into this configuration file.
Makefiles and shell scripts were updated to read this configuration file.